[cargo-verus-toolchains] refactor: expose versioning affordances
